Gemini (#33)
* Revert "Revert "Tries Gemini Pro (#32)""

This reverts commit 2a6d0098ba3dda512a5fbe44013586d9bba91894.

* Try Gemini again
