+++ Trunk-Transcribe/app.go
@@ -27,9 +27,11 @@ import (
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
 	"github.com/aws/aws-sdk-go/service/s3/s3manager"
-	lru "github.com/hashicorp/golang-lru/v2"	
+	"github.com/google/generative-ai-go/genai"
+	lru "github.com/hashicorp/golang-lru/v2"
 	"github.com/slack-go/slack"
 	"golang.org/x/sync/errgroup"
+	"google.golang.org/api/option"
 )
 
 type SlackChannelID string
@@ -61,7 +63,7 @@ var (
 	versusRegex  = regexp.MustCompile(modeString + `.+(vs|versus|verses)(\.)?.+` + modeString)
 	streets      = []string{"Acton", "Ada", "Addison", "Adeline", "Alcatraz", "Allston", "Ashby", "Bancroft", "Benvenue", "Berryman", "Blake", "Bonar", "Bonita", "Bowditch", "Buena", "California", "Camelia", "Carleton", "Carlotta", "Cedar", "Center", "Channing", "Chestnut", "Claremont", "Codornices", "College", "Cragmont", "Delaware", "Derby", "Dwight", "Eastshore", "Edith", "Elmwood", "Euclid", "Francisco", "Fresno", "Gilman", "Grizzly", "Harrison", "Hearst", "Heinz", "Henry", "Hillegass", "Holly", "Hopkins", "Josephine", "Kains", "King", "LeConte", "LeRoy", "Hilgard", "Mabel", "Marin", "Martin", "MLK", "Milvia", "Monterey", "Napa", "Neilson", "Oregon", "Parker", "Piedmont", "Posen", "Rose", "Russell", "Sacramento", "San Pablo", "Santa", "Fe", "Shattuck", "Solano", "Sonoma", "Spruce", "Telegraph", "Alameda", "Thousand", "Oaks", "University", "Vine", "Virginia", "Ward", "Woolsey"}
 	modifiers    = []string{"street", "boulevard", "road", "path", "way", "avenue", "highway"}
-	terms        = []string{"bike", "bicycle", "pedestrian", "vehicle", "injury", "victim", "versus", "transport", "concious", "breathing", "alta bates", "highland", "BFD", "Adam", "ID tech", "ring on three", "code 4", "code 34", "en route", "case number", "berry brothers", "rita run", "DBF", "Flock camera", "10-four", "10-4", "10 four", "His Lordships", "Cesar Chavez Park", "10-9 your traffic", "copy", "tow"}
+	terms        = []string{"bike", "bicycle", "pedestrian", "vehicle", "injury", "victim", "versus", "transport", "concious", "breathing", "alta bates", "highland", "BFD", "Adam", "ID tech", "ring on three", "code 4", "code 34", "en route", "case number", "berry brothers", "rita run", "DBF", "Flock camera", "10-four", "10-4", "10 four", "His Lordships", "Cesar Chavez Park", "10-9 your traffic", "copy", "tow", "kill the beeper"}
 
 	defaultChannelID = BERKELEY // #scanner-dispatches
 
@@ -137,7 +139,7 @@ var (
 
 var notifsMap = map[SlackUserID]Notifs{
 	EMILIE: Notifs{
-		Include:  []string{"auto ped", "auto-ped", "autoped","autobike", "auto bike", "auto bicycle", "auto-bike", "auto-bicycle", "hit and run", "1071", "GSW", "loud reports", "211", "highland", "catalytic", "apple", "261", "code 3", "10-15", "beeper", "1053", "1054", "1055", "1080", "1199", "DBF", "Code 33", "1180", "215", "220", "243", "244", "243", "288", "451", "288A", "243", "207", "212.5", "1079", "1067", "accident", "collision", "fled", "homicide", "fait", "fate", "injuries", "conscious", "responsive", "shooting", "shoot", "coroner", "weapon", "weapons", "gun", "flock", "spikes", "challenging", "beeper", "cage", "tom", "register", "1033 frank", "1033f", "1033", "10-33 frank"},
+		Include:  []string{"auto ped", "auto-ped", "autoped", "autobike", "auto bike", "auto bicycle", "auto-bike", "auto-bicycle", "hit and run", "1071", "GSW", "loud reports", "211", "highland", "catalytic", "apple", "261", "code 3", "10-15", "beeper", "1053", "1054", "1055", "1080", "1199", "DBF", "Code 33", "1180", "215", "220", "243", "244", "243", "288", "451", "288A", "243", "207", "212.5", "1079", "1067", "accident", "collision", "fled", "homicide", "fait", "fate", "injuries", "conscious", "responsive", "shooting", "shoot", "coroner", "weapon", "weapons", "gun", "flock", "spikes", "challenging", "beeper", "cage", "tom", "register", "1033 frank", "1033f", "1033", "10-33 frank"},
 		NotRegex: regexp.MustCompile("no (weapon|gun)s?"),
 		Regex:    versusRegex,
 		Channels: []SlackChannelID{BERKELEY, UCPD},
@@ -148,7 +150,7 @@ var notifsMap = map[SlackUserID]Notifs{
 		Channels: []SlackChannelID{BERKELEY, UCPD, ALBANY, EMERYVILLE},
 	},
 	MARC: Notifs{
-		Include: []string{"hit and run", "autobike", "auto bike", "auto bicycle", "auto bicyclist", "auto ped", "auto-ped", "autoped"},
+		Include:  []string{"hit and run", "autobike", "auto bike", "auto bicycle", "auto bicyclist", "auto ped", "auto-ped", "autoped"},
 		Regex:    versusRegex,
 		Channels: []SlackChannelID{BERKELEY, UCPD},
 	},
@@ -163,7 +165,7 @@ var notifsMap = map[SlackUserID]Notifs{
 		Channels: []SlackChannelID{BERKELEY, UCPD},
 	},
 	HELEN: Notifs{
-		Include: []string{"hit and run", "autobike", "auto bike", "auto bicycle", "auto bicyclist", "auto ped", "auto-ped", "autoped", "marin", "hopkins"},
+		Include:  []string{"hit and run", "autobike", "auto bike", "auto bicycle", "auto bicyclist", "auto ped", "auto-ped", "autoped", "marin", "hopkins"},
 		Regex:    versusRegex,
 		Channels: []SlackChannelID{BERKELEY},
 	},
@@ -172,29 +174,31 @@ var notifsMap = map[SlackUserID]Notifs{
 // cloudflare setup
 var cloudflareApiToken string = os.Getenv("CLOUDFLARE_API_TOKEN")
 var cloudflareAccountID string = os.Getenv("CLOUDFLARE_ACCOUNT_ID")
-var cloudflareWhisperUrl string = "https://api.cloudflare.com/client/v4/accounts/"+cloudflareAccountID+"/ai/run/@cf/openai/whisper-large-v3-turbo"
+var cloudflareWhisperUrl string = "https://api.cloudflare.com/client/v4/accounts/" + cloudflareAccountID + "/ai/run/@cf/openai/whisper-large-v3-turbo"
 var r2Key string = os.Getenv("CLOUDFLARE_R2_KEY")
 var r2Secret string = os.Getenv("CLOUDFLARE_R2_SECRET")
+
+// google gemini
+var geminiApiKey string = os.Getenv("GEMINI_API_KEY")
 
 // slack setup
 var webhookUrl string = os.Getenv("SLACK_WEBHOOK_URL")
 var webhookUrlUCPD string = os.Getenv("SLACK_WEBHOOK_URL_UCPD")
-var slackapiSecret string = os.Getenv("SLACK_API_SECRET")	
+var slackapiSecret string = os.Getenv("SLACK_API_SECRET")
 
 //go:embed templates/*
 var resources embed.FS
 
 var t = template.Must(template.ParseFS(resources, "templates/*"))
 
-type Config struct {	
-	uploader       *s3manager.Uploader	
+type Config struct {
+	uploader       *s3manager.Uploader
 	slackClient    *slack.Client
 	webhookUrl     string
 	webhookUrlUCPD string
 }
 
 var dedupeCache *lru.Cache[string, bool]
-
 
 func init() {
 	var err error
@@ -215,27 +219,27 @@ func main() {
 		port = "8080"
 
 	}
-	
+
 	var api *slack.Client
 	if slackapiSecret == "" || webhookUrl == "" {
 		log.Println("Missing SLACK_API_SECRET or SLACK_WEBHOOK_URL. Slack notifications disabled.")
 	} else {
 		api = slack.New(slackapiSecret)
-	}	
-	
-        // R2 setup
+	}
+
+	// R2 setup
 	endpoint := fmt.Sprintf("https://%s.r2.cloudflarestorage.com", cloudflareAccountID)
 	fmt.Println("Using cloudflare R2 endpoint: ", endpoint)
-	
+
 	r2Config := &aws.Config{
 		Region:      aws.String("auto"),
 		Credentials: credentials.NewStaticCredentials(r2Key, r2Secret, ""),
-		Endpoint: aws.String(endpoint),		
-	}	
+		Endpoint:    aws.String(endpoint),
+	}
 	uploader := s3manager.NewUploader(session.New(r2Config))
-	
-	config := &Config{		
-		uploader:       uploader,		
+
+	config := &Config{
+		uploader:       uploader,
 		slackClient:    api,
 		webhookUrl:     webhookUrl,
 		webhookUrlUCPD: webhookUrlUCPD,
@@ -373,13 +377,13 @@ func dedupeDispatch(meta Metadata) (dupe bool) {
 	for _, src := range meta.SrcList {
 		srcs += fmt.Sprintf(".%v", src.Src)
 	}
-	
+
 	startTime := meta.StartTime - (meta.StartTime % 5) // time to the nearest 5 second increment
 	dedupeKey := fmt.Sprintf("tg.%d.start.%d.srcs%s", meta.Talkgroup, startTime, srcs)
 
 	// atomically check-or-set. Return whether the key already existed.
 	exists, _ := dedupeCache.ContainsOrAdd(dedupeKey, true)
-	
+
 	return exists
 }
 
@@ -387,6 +391,7 @@ func dedupeDispatch(meta Metadata) (dupe bool) {
 func transcribeAndUpload(ctx context.Context, config *Config, key string, data []byte, metadata Metadata) (string, error) {
 
 	msg, err := whisper(ctx, data)
+
 	if err == nil {
 		fmt.Println(key+": ", msg)
 	} else {
@@ -396,13 +401,20 @@ func transcribeAndUpload(ctx context.Context, config *Config, key string, data [
 	metadata.AudioText = msg
 	metadata.URL = fmt.Sprintf("https://trunk-transcribe.fly.dev/audio?link=%s", key)
 
-	wg, gctx := errgroup.WithContext(ctx)	
+	gmsg, gerr := gemini(ctx, data)
+	if gerr != nil {
+		fmt.Println("Error with gemini: ", gerr)
+	} else {
+		fmt.Println("Gemini response: ", gmsg)
+	}
+
+	wg, gctx := errgroup.WithContext(ctx)
 
 	//upload to Cloudflare R2 (with s3 compatible api)
-	wg.Go(func() error { 
+	wg.Go(func() error {
 		return uploadS3(gctx, config.uploader, key, bytes.NewReader(data), metadata)
 	})
-	
+
 	wg.Go(func() error {
 		return postToSlack(gctx, config, key, bytes.NewReader(data), metadata)
 	})
@@ -410,16 +422,70 @@ func transcribeAndUpload(ctx context.Context, config *Config, key string, data [
 	return msg, err
 }
 
+func gemini(ctx context.Context, data []byte) (string, error) {
+
+	// remove silence
+	// segment, err := gaudio.LoadAudio(bytes.NewReader(data), gaudio.FormatWAVE)
+	// if err != nil {
+	// 	return "", err
+	// }
+
+	// // Get leading and trailing silence
+	// segment.RemoveStartAndEndSilence(threshold)
+	// buffer := new(bytes.Buffer)
+	// err = segment.Export(buffer, gaudio.FormatWAVE)
+	// if err != nil {
+	// 	return "", err
+	// }
+
+	client, err := genai.NewClient(ctx, option.WithAPIKey(geminiApiKey))
+	if err != nil {
+		return "", err
+	}
+
+	prompt := strings.Join(append(streets, append(modifiers, terms...)...), ", ")
+	parts := []genai.Part{
+		genai.Blob{MIMEType: "audio/mp3", Data: data},
+		genai.Text("Please transcribe the audio. " + prompt),
+	}
+
+	model := client.GenerativeModel("gemini-1.5-pro")
+	// Generate content using the prompt.
+	resp, err := model.GenerateContent(ctx, parts...)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	// Handle the response of generated text
+
+	b, _ := json.MarshalIndent(resp, "", " ")
+	fmt.Println(string(b))
+
+	var buf strings.Builder
+
+	for _, c := range resp.Candidates {
+		if c.Content != nil {
+			for _, part := range c.Content.Parts {
+				fmt.Fprintf(&buf, "%v", part)
+				buf.WriteString("\n")
+			}
+		}
+	}
+
+	fmt.Println(buf.String())
+	return buf.String(), nil
+}
+
 // whisper transcribes the audio with cloudflare Whisper
 func whisper(ctx context.Context, data []byte) (string, error) {
-	prompt := strings.Join(append(streets, append(modifiers, terms...)...), ", ")	
-	
+	prompt := strings.Join(append(streets, append(modifiers, terms...)...), ", ")
+
 	enc := base64.StdEncoding.EncodeToString(data)
-	payload, err := json.Marshal(CloudflareWhisperInput {
-		Audio: enc,
+	payload, err := json.Marshal(CloudflareWhisperInput{
+		Audio:  enc,
 		Prompt: prompt,
 	})
-	
+
 	if err != nil {
 		return "", err
 	}
@@ -427,8 +493,8 @@ func whisper(ctx context.Context, data []byte) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	req.Header.Set("Authorization", "Bearer " + cloudflareApiToken)
-		
+	req.Header.Set("Authorization", "Bearer "+cloudflareApiToken)
+
 	resp, err := http.DefaultClient.Do(req)
 	if err != nil {
 		fmt.Printf("Error calling cloudflare: %v\n", err)
@@ -445,7 +511,7 @@ func whisper(ctx context.Context, data []byte) (string, error) {
 		return "", err
 	}
 	fmt.Println("Response from cloudflare: ", output)
-	return output.Result.Text, nil	
+	return output.Result.Text, nil
 }
 
 // transcribeAndUpload uploads the audio to S3
@@ -611,159 +677,4 @@ func ExtractSlackMeta(meta Metadata, channelID SlackChannelID, notifsMap map[Sla
 func writeErr(w http.ResponseWriter, err error) {
 	fmt.Println("Error: ", err.Error())
 	http.Error(w, err.Error(), http.StatusInternalServerError)
-}
-
-// Structures to parse json of the form:
-//
-//	{
-//	  "freq": 772693750,
-//	  "start_time": 1702513859,
-//	  "stop_time": 1702513867,
-//	  "emergency": 0,
-//	  "priority": 4,
-//	  "mode": 0,
-//	  "duplex": 0,
-//	  "encrypted": 0,
-//	  "call_length": 6,
-//	  "talkgroup": 3105,
-//	  "talkgroup_tag": "Berkeley PD1",
-//	  "talkgroup_description": "Police Dispatch",
-//	  "talkgroup_group_tag": "Law Dispatch",
-//	  "talkgroup_group": "Berkeley",
-//	  "audio_type": "digital",
-//	  "short_name": "Berkeley",
-//	  "freqList": [
-//	    {
-//	      "freq": 772693750,
-//	      "time": 1702513859,
-//	      "pos": 0,
-//	      "len": 1.08,
-//	      "error_count": "100",
-//	      "spike_count": "7"
-//	    },
-//	    {
-//	      "freq": 772693750,
-//	      "time": 1702513862,
-//	      "pos": 1.08,
-//	      "len": 3.78,
-//	      "error_count": "21",
-//	      "spike_count": "1"
-//	    },
-//	  ],
-//	  "srcList": [
-//	    {
-//	      "src": 3113003,
-//	      "time": 1702513859,
-//	      "pos": 0,
-//	      "emergency": 0,
-//	      "signal_system": "",
-//	      "tag": "Dispatch"
-//	    },
-//	    {
-//	      "src": 3124119,
-//	      "time": 1702513862,
-//	      "pos": 1.08,
-//	      "emergency": 0,
-//	      "signal_system": "",
-//	      "tag": ""
-//	    },
-//	  ]
-//	}
-type Frequency struct {
-	Freq int64   `json:"freq,omitempty"`
-	Time int64   `json:"time,omitempty"`
-	Pos  float64 `json:"pos,omitempty"`
-}
-type Source struct {
-	Src          int64   `json:"src,omitempty"`
-	Time         int64   `json:"time,omitempty"`
-	Pos          float64 `json:"pos,omitempty"`
-	Emergency    int64   `json:"emergency,omitempty"`
-	SignalSystem string  `json:"signal_system,omitempty"`
-	Tag          string  `json:"tag,omitempty"`
-}
-
-type Metadata struct {
-	Freq              int64       `json:"freq,omitempty"`
-	StartTime         int64       `json:"start_time,omitempty"`
-	StopTime          int64       `json:"stop_time,omitempty"`
-	Emergency         int64       `json:"emergency,omitempty"`
-	Priority          int64       `json:"priority,omitempty"`
-	Mode              int64       `json:"mode,omitempty"`
-	Duplex            int64       `json:"duplex,omitempty"`
-	Encrypted         int64       `json:"encrypted,omitempty"`
-	CallLength        int64       `json:"call_length,omitempty"`
-	Talkgroup         int64       `json:"talkgroup,omitempty"`
-	TalkgroupTag      string      `json:"talkgroup_tag,omitempty"`
-	TalkGroupDesc     string      `json:"talkgroup_description,omitempty"`
-	TalkGroupGroupTag string      `json:"talkgroup_group_tag,omitempty"`
-	TalkGroupGroup    string      `json:"talkgroup_group,omitempty"`
-	AudioType         string      `json:"audio_type,omitempty"`
-	ShortName         string      `json:"short_name,omitempty"`
-	AudioText         string      `json:"audio_text,omitempty"`
-	URL               string      `json:"url,omitempty"`
-	SrcList           []Source    `json:"srcList,omitempty"`
-	FreqList          []Frequency `json:"freqList,omitempty"`
-}
-
-type Notifs struct {
-	Include  []string
-	Regex    *regexp.Regexp
-	NotRegex *regexp.Regexp
-	Channels []SlackChannelID //the channels to listen on
-}
-
-type SlackMeta struct {
-	Mentions []string `json:"mentions,omitempty"`
-	Address  Address  `json:"address,omitempty"`
-}
-
-// Address struct encapsulates address info to extract from transcription text
-// Example transcriptions: "2605 Durant", "Can you start for Russell and California please?"
-type Address struct {
-	City           string   `json:"city,omitempty"`
-	PrimaryAddress string   `json:"primary,omitempty"`
-	Streets        []string `json:"streets,omitempty"`
-}
-
-func (addr Address) AppendStreet(street string) Address {
-	for _, st := range addr.Streets {
-		if st == street {
-			return addr
-		}
-	}
-	addr.Streets = append(addr.Streets, street)
-	return addr
-}
-
-func (addr Address) String() string {
-	switch {
-	case addr.PrimaryAddress != "":
-		return addr.PrimaryAddress
-	case len(addr.Streets) > 1:
-		return addr.Streets[0] + " and " + addr.Streets[1]
-	default:
-		return ""
-	}
-}
-
-
-type TranscriptionInfo struct {
-	Language string  `json:"language,omitempty"`
-	Duration string  `json:"duration,omitempty"`
-	Text     string  `json:"text,omitempty"`
-	WordCount int    `json:"word_count,omitempty"`
-}
-
-type CloudflareWhisperInput struct {
-	Audio string 	`json:"audio,omitempty"`
-	Prompt string 	`json:"initial_prompt,omitempty"`
-	Prefix string   `json:"prefix,omitempty"`
-}
-
-type CloudflareWhisperOutput struct {
-	Result TranscriptionInfo `json:"result,omitempty"`
-	Success bool             `json:"success,omitempty"`
-	Errors  []string         `json:"errors,omitempty"`
-	Messages []string        `json:"messages,omitempty"`
 }
